Создание проекта
Настроен pom.xml
#15 - уточнён Seeker.java.
Обновлены зависимости.
