+++ .gitignore
@@ -15,3 +15,7 @@ hs_err_pid*
 target/*
 
 *.xls
+
+*.idea*
+
+*.iml
