+++ client-exel/pom.xml
@@ -24,6 +24,11 @@
             <version>4.11</version>
             <scope>test</scope>
         </dependency>
+		<dependency>
+			<groupId>ru.gosha</groupId>
+            <artifactId>ru.gosha.server-client</artifactId>
+            <version>0.1</version>
+		</dependency>
     </dependencies>
 
     <properties>
