--- client-exel/src/main/java/ru/gosha/SeekerType.java
@@ -1,5 +1,0 @@
-package ru.gosha;
-
-public enum SeekerType {
-    Teacher, StudyGroup
-}
