+++ serverClient/pom.xml
@@ -5,14 +5,17 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>ru.gosha</groupId>
-    <artifactId>untitled</artifactId>
-    <packaging>pom</packaging>
-    <version>1.0-SNAPSHOT</version>
-    <modules>
-        <module>server-exel</module>
-        <module>client-exel</module>
-		<module>serverClient</module>
-    </modules>
+    <artifactId>ru.gosha.serverClient</artifactId>
+    <packaging>jar</packaging>
+    <version>0.1</version>
+    <dependencies>
+        <dependency>
+            <groupId>junit</groupId>
+            <artifactId>junit</artifactId>
+            <version>4.11</version>
+            <scope>test</scope>
+        </dependency>
+    </dependencies>
 
 
     <properties>
