+++ serverClient/src/main/java/ru/gosha/serverClient/SeekerType.java
@@ -1,0 +1,12 @@
+package ru.gosha.serverClient;
+
+public enum SeekerType {
+    /**
+     * Искатель-преподаватель.
+     */
+    Teacher,
+    /**
+     * Искатель-студент группы.
+     */
+    StudyGroup
+}
