+++ serverClient/src/test/java/testSeeker.java
@@ -1,0 +1,18 @@
+
+import ru.gosha.serverClient.*;
+
+import java.time.LocalDate;
+import java.time.ZoneId;
+import java.util.Date;
+import java.util.SimpleTimeZone;
+import org.junit.Test;
+
+public class testSeeker {
+
+    @Test
+    public void startTestSeeker() {
+        //Seeker test = new Seeker("name", SeekerType.Teacher, new Date(), LocalDate., new SimpleTimeZone(+3 * 60 * 60 * 1000, "Moscow"), "Москва, проспект Вернадского, 78, РТУ МИРЭА");
+
+
+    }
+}
